+++ gui_qt/explorerchunkterrain.cpp
@@ -4,17 +4,17 @@
 #include "baseexplorerchunk.h"
 #include "../lib_paysages/camera.h"
 
-ExplorerChunkTerrain::ExplorerChunkTerrain(Renderer* renderer, float x, float z, float size, int nbchunks) : BaseExplorerChunk(renderer)
+ExplorerChunkTerrain::ExplorerChunkTerrain(Renderer* renderer, double x, double z, double size, int nbchunks) : BaseExplorerChunk(renderer)
 {
     _startx = x;
     _startz = z;
     _size = size;
-    _overall_step = size * (float)nbchunks;
+    _overall_step = size * (double)nbchunks;
     
     _tessellation_max_size = 32;
-    _tessellation = new float[(_tessellation_max_size + 1) * (_tessellation_max_size + 1)];
+    _tessellation = new double[(_tessellation_max_size + 1) * (_tessellation_max_size + 1)];
     _tessellation_current_size = 0;
-    _tessellation_step = _size / (float)_tessellation_max_size;
+    _tessellation_step = _size / (double)_tessellation_max_size;
     
     setMaxTextureSize(128);
     
@@ -49,7 +49,7 @@ bool ExplorerChunkTerrain::onMaintainEvent()
             {
                 if (_tessellation_current_size == 0 || i % old_tessellation_inc != 0 || j % old_tessellation_inc != 0)
                 {
-                    float height = renderer->getTerrainHeight(renderer, _startx + _tessellation_step * (float)i, _startz + _tessellation_step * (float)j);
+                    double height = renderer->getTerrainHeight(renderer, _startx + _tessellation_step * (double)i, _startz + _tessellation_step * (double)j);
                     _tessellation[j * (_tessellation_max_size + 1) + i] = height;
                 }
             }
@@ -98,7 +98,7 @@ void ExplorerChunkTerrain::onRenderEvent(QGLWidget* widget)
 {
     _lock_data.lock();
     int tessellation_size = _tessellation_current_size;
-    float tsize = 1.0 / (float)_tessellation_max_size;
+    double tsize = 1.0 / (double)_tessellation_max_size;
     _lock_data.unlock();
     
     if (tessellation_size <= 1)
@@ -106,24 +106,24 @@ void ExplorerChunkTerrain::onRenderEvent(QGLWidget* widget)
         return;
     }
 
-    int tessellation_inc = _tessellation_max_size / (float)tessellation_size;
+    int tessellation_inc = _tessellation_max_size / (double)tessellation_size;
     for (int j = 0; j < _tessellation_max_size; j += tessellation_inc)
     {
         glBegin(GL_QUAD_STRIP);
         for (int i = 0; i <= _tessellation_max_size; i += tessellation_inc)
         {
-            glTexCoord2d(tsize * (float)i, tsize * (float)j);
-            glVertex3d(_startx + _tessellation_step * (float)i, _tessellation[j * (_tessellation_max_size + 1) + i], _startz + _tessellation_step * (float)j);
-            glTexCoord2d(tsize * (float)i, tsize * (float)(j + tessellation_inc));
-            glVertex3d(_startx + _tessellation_step * (float)i, _tessellation[(j + tessellation_inc) * (_tessellation_max_size + 1) + i], _startz + _tessellation_step * (float)(j + tessellation_inc));
+            glTexCoord2d(tsize * (double)i, tsize * (double)j);
+            glVertex3d(_startx + _tessellation_step * (double)i, _tessellation[j * (_tessellation_max_size + 1) + i], _startz + _tessellation_step * (double)j);
+            glTexCoord2d(tsize * (double)i, tsize * (double)(j + tessellation_inc));
+            glVertex3d(_startx + _tessellation_step * (double)i, _tessellation[(j + tessellation_inc) * (_tessellation_max_size + 1) + i], _startz + _tessellation_step * (double)(j + tessellation_inc));
         }
         glEnd();
     }
 }
 
-float ExplorerChunkTerrain::getDisplayedSizeHint(CameraDefinition* camera)
+double ExplorerChunkTerrain::getDisplayedSizeHint(CameraDefinition* camera)
 {
-    float distance;
+    double distance;
     Vector3 center;
 
     center = getCenter();
@@ -140,7 +140,7 @@ float ExplorerChunkTerrain::getDisplayedSizeHint(CameraDefinition* camera)
     }
 }
 
-Color ExplorerChunkTerrain::getTextureColor(float x, float y)
+Color ExplorerChunkTerrain::getTextureColor(double x, double y)
 {
     Vector3 location = {_startx + x * _size, 0.0, _startz + y * _size};
     return renderer()->applyTextures(renderer(), location, 0.01);
